+++ packages/active_grid_core/lib/web_auth_enabler/configure_nonweb.dart
@@ -1,0 +1,2 @@
+/// Stub Function used for non web platforms
+Future<void> enableWebAuth() async {}