+++ packages/active_grid_form/test/component_test.dart
@@ -12,7 +12,7 @@ import 'common.dart';
 
 void main() {
   setUpAll(() {
-    registerFallbackValue<FormData>(FormData("Title", [], [], {}));
+    registerFallbackValue<FormData>(FormData('Title', [], [], {}));
   });
 
   group('Text', () {