+++ packages/active_grid_core/CHANGELOG.md
@@ -1,2 +1,12 @@
+## [0.1.1+2]
+* Fix closeWebView invocation on Web
+
+## [0.1.1+1]
+* Cleanup
+* Add Authentication Documentation to README
+
+## [0.1.1]
+* Fix Web Authentication
+
 ## [0.1.0]
 * Initial Release